requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
ebooklib
beautifulsoup4
requests
PyPDF2
openai==0.27.8